% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_server.R
\name{barplot_server}
\alias{barplot_server}
\title{Barplot Server}
\usage{
barplot_server(
  id,
  sample_data_function,
  feature_data = shiny::reactive(NULL),
  group_data = shiny::reactive(NULL),
  barplot_xlab = shiny::reactive(""),
  barplot_ylab = shiny::reactive(""),
  barplot_title = shiny::reactive(""),
  barplot_label = shiny::reactive("Feature"),
  drilldown = shiny::reactive(F),
  ...
)
}
\arguments{
\item{id}{Module ID}

\item{sample_data_function}{A shiny::reactive that returns a function.
The function must take an argument called ".feature" and return a
dataframe with columns "sample_name", "feature_name", "group_name", and
"feature_value".}

\item{feature_data}{A shiny::reactive that returns a dataframe with columns
"feature_name", and optionally "feature_display" and "feature_class".
 Each value in the "feature_name" column should only appear once.}

\item{group_data}{A shiny::reactive that returns a dataframe with columns
"group_name", "group_display", and optionally "group_description" and
"group_color". Each value in the "group_name" column should only appear once.}

\item{barplot_xlab}{A shiny::reactive that returns a string}

\item{barplot_ylab}{A shiny::reactive that returns a string}

\item{barplot_title}{A shiny::reactive that returns a string}

\item{barplot_label}{A shiny::reactive that returns a string}

\item{drilldown}{A shiny::reactive that returns True or False}

\item{...}{shiny::reactives passed to barplot_server}
}
\description{
Barplot Server
}
