% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_server2.R
\name{barplot_server2}
\alias{barplot_server2}
\title{Barplot Server2}
\usage{
barplot_server2(
  id,
  barplot_data,
  group_data = shiny::reactive(NULL),
  barplot_xlab = shiny::reactive(""),
  barplot_ylab = shiny::reactive(""),
  barplot_title = shiny::reactive(""),
  barplot_label = shiny::reactive("Feature"),
  drilldown = shiny::reactive(F),
  ...
)
}
\arguments{
\item{id}{Module ID}

\item{barplot_data}{A shiny::reactive that returns a dataframe with
columns "sample_name", "group_name", feature_display", and "feature_value".}

\item{group_data}{A shiny::reactive that returns a dataframe with columns
group_display", and optionally, "group_description". Each value in the
"group_name" column should only appear once.}

\item{barplot_xlab}{A shiny::reactive that returns a string}

\item{barplot_ylab}{A shiny::reactive that returns a string}

\item{barplot_title}{A shiny::reactive that returns a string}

\item{barplot_label}{A shiny::reactive that returns a string}

\item{drilldown}{A shiny::reactive that returns True or False}

\item{...}{shiny::reactives passed to drilldown_scatterplot_server}
}
\description{
Barplot Server2
}
