% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_ui2.R
\name{barplot_ui2}
\alias{barplot_ui2}
\title{Barplot UI2}
\usage{
barplot_ui2(id, ...)
}
\arguments{
\item{id}{Module ID}

\item{...}{Arguments passed to drilldown_scatterplot_ui}
}
\description{
Barplot UI2
}
