% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_plot_server.R
\name{distributions_plot_server}
\alias{distributions_plot_server}
\title{Distributions Plot Server}
\usage{
distributions_plot_server(
  id,
  sample_data_function,
  feature_data = shiny::reactive(NULL),
  group_data = shiny::reactive(NULL),
  distplot_xlab = shiny::reactive(""),
  distplot_title = shiny::reactive(NULL),
  scale_method_default = shiny::reactive("None"),
  feature_default = shiny::reactive(NULL),
  drilldown = shiny::reactive(F),
  ...
)
}
\arguments{
\item{id}{Module ID}

\item{sample_data_function}{A shiny::reactive that returns a function.
The function must take an argument called ".feature" and return a
dataframe with columns "sample_name", "feature_name", "group_name", and
"feature_value",}

\item{feature_data}{A shiny::reactive that returns a dataframe with columns
"feature_name", and "feature_display". Any other additional columns will be
 names of classes to group the features by. Each value in the "feature_name"
 column should only appear once.}

\item{group_data}{A shiny::reactive that returns a dataframe with columns
"group_name", "group_display", and optionally "group_description" and
"group_color". Each value in the "group_name"column should only appear once.}

\item{distplot_xlab}{A shiny::reactive that returns a string}

\item{distplot_title}{A shiny::reactive that returns a string}

\item{scale_method_default}{A shiny::reactive that returns a string}

\item{feature_default}{A shiny::reactive that returns a string}

\item{drilldown}{A shiny::reactive that returns True or False}

\item{...}{shiny::reactives passed to drilldown_histogram_server}
}
\description{
Distributions Plot Server
}
