% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_plot_server2.R
\name{distributions_plot_server2}
\alias{distributions_plot_server2}
\title{Distributions Plot Server}
\usage{
distributions_plot_server2(
  id,
  distplot_data,
  group_data = shiny::reactive(NULL),
  plot_type = shiny::reactive("Violin"),
  distplot_xlab = shiny::reactive(""),
  distplot_title = shiny::reactive(NULL),
  drilldown = shiny::reactive(F),
  ...
)
}
\arguments{
\item{id}{Module ID}

\item{distplot_data}{A shiny::reactive that returns a dataframe with
columns "sample_name", "group_name", feature_display", and "feature_value".}

\item{group_data}{A shiny::reactive that returns a dataframe with columns
"group_name", "group_display", and optionally "group_description" and
"group_color". Each value in the "group_name"column should only appear once.}

\item{plot_type}{A shiny::reactive that returns a string, eith "Violin" or
"Box"}

\item{distplot_xlab}{A shiny::reactive that returns a string}

\item{distplot_title}{A shiny::reactive that returns a string}

\item{drilldown}{A shiny::reactive that returns True or False}

\item{...}{shiny::reactives passed to drilldown_histogram_server}
}
\description{
Distributions Plot Server
}
