% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_plot_ui2.R
\name{distributions_plot_ui2}
\alias{distributions_plot_ui2}
\title{Distributions Plot UI2}
\usage{
distributions_plot_ui2(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
Distributions Plot UI2
}
