% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drilldown_scatterplot_server.R
\name{drilldown_scatterplot_server}
\alias{drilldown_scatterplot_server}
\title{Drilldown Scatterplot Server}
\usage{
drilldown_scatterplot_server(
  id,
  scatterplot_data,
  x_feature_input = NULL,
  y_feature_input = NULL,
  selected_group = shiny::reactive("Clicked Group")
)
}
\arguments{
\item{id}{Module ID}

\item{scatterplot_data}{A shiny::reactive that returns a dataframe with columns
"sample", "group", "feature", "feature_value"}

\item{x_feature_input}{A shiny::reactive that returns a string}

\item{y_feature_input}{A shiny::reactive that returns a string}

\item{selected_group}{A string, this gets added to the sample label}
}
\description{
Drilldown Scatterplot Server
}
