% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_server2.R
\name{heatmap_server2}
\alias{heatmap_server2}
\title{Heatmap Server}
\usage{
heatmap_server2(
  id,
  heatmap_data,
  group_data = shiny::reactive(NULL),
  summarise_function = shiny::reactive(stats::cor),
  drilldown = shiny::reactive(F),
  mock_event_data = shiny::reactive(NULL),
  ...
)
}
\arguments{
\item{id}{Module ID}

\item{heatmap_data}{A shiny::reactive that returns a dataframe with a columns
named "sample_name", "group_name", "dataset_name", "feature_value",
"feature_display", "response_display", "response_value", and optionally
"feature_order".}

\item{group_data}{A shiny::reactive that returns a dataframe with columns
"group_display", and optionally "group_description" and
"group_order". Each value in the "group_display" column should only appear
once.}

\item{summarise_function}{A shiny::reactive that returns a function. The
function must take two vectors. The first one will be
the "feature_value" column of heatmap_data, and the second will be
the "response_value" of heatmap_data. The function must return one
numeric value.}

\item{drilldown}{A shiny::reactive that returns True or False}

\item{mock_event_data}{A shiny::reactive that returns a dataframe. For
testing purposes only. Must have columns "curveNumber", "pointNumber", "x",
"y", and "z". The "x" column corresponds to the group selected, and the
"y" column corresponds to feature selected.}

\item{...}{arguments sent to plotly_scatter}
}
\description{
Heatmap Server
}
