% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_ui.R
\name{heatmap_ui}
\alias{heatmap_ui}
\title{Heatmap UI}
\usage{
heatmap_ui(
  id,
  title = "",
  html = htmltools::includeMarkdown(get_markdown_path("distplot1")),
  ...
)
}
\arguments{
\item{id}{Module ID}

\item{title}{A string}

\item{html}{A string that is HTML}

\item{...}{Arguments passed to drilldown_scatterplot_ui}
}
\description{
Heatmap UI
}
