% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_ui2.R
\name{heatmap_ui2}
\alias{heatmap_ui2}
\title{Heatmap UI2}
\usage{
heatmap_ui2(id, ...)
}
\arguments{
\item{id}{Module ID}

\item{...}{Arguments passed to drilldown_scatterplot_ui}
}
\description{
Heatmap UI2
}
