% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_items.R
\name{imgLinkBox}
\alias{imgLinkBox}
\title{Img Link Box}
\usage{
imgLinkBox(..., linkId, title, imgSrc, boxText, linkText)
}
\arguments{
\item{...}{arguments to shinydashboard::box}

\item{linkId}{a string}

\item{title}{a string}

\item{imgSrc}{a string}

\item{boxText}{a string}

\item{linkText}{a string}
}
\description{
Img Link Box
}
