% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{refactor_by_tbl_value_column}
\alias{refactor_by_tbl_value_column}
\title{Refactor By Tibble Value Column}
\usage{
refactor_by_tbl_value_column(tbl, reorder_method = "None")
}
\arguments{
\item{tbl}{A Tibble with columns "feature_value", "group_display"}

\item{reorder_method}{One of "None", "Median", "Mean", "Max", "Min}
}
\description{
Refactor By Tibble Value Column
}
