% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_plot_server2.R
\name{distributions_plot_server2}
\alias{distributions_plot_server2}
\title{Distributions Plot Server}
\usage{
distributions_plot_server2(
  id,
  distplot_data,
  group_data = shiny::reactive(NULL),
  dataset_data = shiny::reactive(NULL),
  plot_type = shiny::reactive("Violin"),
  scale_method = shiny::reactive("None"),
  reorder_method = shiny::reactive("None"),
  distplot_xlab = shiny::reactive(""),
  drilldown = shiny::reactive(F),
  mock_event_data = shiny::reactive(NULL),
  ...
)
}
\arguments{
\item{id}{Module ID}

\item{distplot_data}{A shiny::reactive that returns a dataframe with
columns "sample_name", "group_name", feature_display", "feature_value", and
"dataset_name". There will be one plot for each value in "dataset_name".}

\item{group_data}{A shiny::reactive that returns a dataframe with columns
"group_name", "group_display", and optionally "group_description" and
"group_color". Each value in the "group_name" column should only appear once.}

\item{dataset_data}{A shiny::reactive that returns a dataframe with columns
"dataset_name", and "dataset_display".}

\item{plot_type}{A shiny::reactive that returns a string, either "Violin" or
"Box"}

\item{scale_method}{A shiny::reactive that returns a string, one of (
"None", Log2", "Log2 + 1", "Log10 + 1", "Log10")}

\item{reorder_method}{A shiny::reactive that returns a string, one of (
"None", "Median", "Mean", "Max", "Min")}

\item{distplot_xlab}{A shiny::reactive that returns a string}

\item{drilldown}{A shiny::reactive that returns True or False}

\item{mock_event_data}{A shiny::reactive that returns a dataframe. For
testing purposes only. Must have columns "curveNumber", "pointNumber", "x",
"y", and "key". The "x" column corresponds to the group selected, and the
"key" column corresponds to dataset selected.}

\item{...}{arguments sent to plotly_histogram}
}
\description{
Distributions Plot Server
}
