% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_server.R
\name{heatmap_server}
\alias{heatmap_server}
\title{Heatmap Server}
\usage{
heatmap_server(
  id,
  feature_classes,
  response_features,
  feature_data_function,
  response_data_function,
  summarise_function_list = shiny::reactive(stats::cor),
  drilldown = shiny::reactive(F),
  default_feature = shiny::reactive(NULL),
  default_class = shiny::reactive(NULL),
  ...
)
}
\arguments{
\item{id}{Module ID}

\item{feature_classes}{A shiny::reactive that returns a list of feature
classes. This will be passed to the choices argument of shiny::selectInput.
The selected class will be passed to the feature_data_function as .class.}

\item{response_features}{A shiny::reactive that returns a list of features.
This will be passed to the choices argument of shiny::selectInput.
The selected feature will be passed to the repsonse_data_function as .feature.}

\item{feature_data_function}{A shiny::reactive that returns a function
The function must take an argument called ".feature_class" and return a
dataframe with columns "sample", "group", "feature", "feature_value",
"feature_order", "group_description", "color"}

\item{response_data_function}{A shiny::reactive that returns a function
The function must take an argument called ".feature" and return a
dataframe with columns "sample", "feature_value"}

\item{summarise_function_list}{A shiny::reactive that returns a either a function
or a named list of functions. If a list is passed, it will be passed to
shiny::selectInput. Each function must take vectors. The first one will be
the "feature_value" column of feature_data_function, and the second will be
the "feature_value" of repsonse_data_function. Each function must return one
numeric value.}

\item{drilldown}{A shiny::reactive that returns True or False}

\item{default_feature}{A shiny::reactive that returns a string that
is one of the values in response_features}

\item{default_class}{A shiny::reactive that returns a string that is one of
the values in feature_classes}

\item{...}{shiny::reactives passed to drilldown_scatterplot_server}
}
\description{
Heatmap Server
}
